refactor: move controller functions to respective route handler files
also simplify server instance's superuser auth by using `autoRefreshThreshold`
